Move RegisterPoseEvent to a subclass of EntityRenderersEvent
